Simplify check in function ad_import.
There is no need to check for the undef value of $data, as well as it is
an hash with at least one key. The idea is to just test for $data being
an hash, and the following loop will not execute if there are no keys in
the hash (avoiding a keys invocation).